Graphics_Algorithms/requirements.txt
glfw==2.10.0
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
